requirement.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1
Django==4.2
django-filter==23.1
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.3
